codex: move exec json output tests back out of impl (#15106)
